requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
firebase-admin
google-cloud-firestore
